django/pyproject.toml
[tool.poetry]
name = "casestudy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2.0"
django-cors-headers = ">=4.2.0"
djangorestframework = ">=3.11,<4.0"
psycopg2-binary = ">=2.8"
redis = ">=4.0"
requests = ">=2.31.0"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
